// healthCare-frontend/src/app/models/Unidade.ts
export interface Unidade {
  id?: number;
  ds_nome: string;
  cnpj: string;
  nr_telefone?: string;
  email: string;
  ieSituacao: string;
  capacidade_atendimento?: string;
  horario_funcionamento?: string;
  tipo: string;
  especialidades_oferecidas?: string;
  nr_cep: string;
  estado?: string;
  cidade?: string;
  bairro?: string;
  endereco?: string;
  nr_endereco?: string;
  complemento?: string;
  como_chegar?: string;
  dt_criacao: string;
}